+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ---
 
+## [5.3.5](https://github.com/fritzlabs/swift-framework/releases/tag/5.3.5)
+
+1. Xcode Version: 11.4.X
+2. Adds support for image segmentation to the Dataset Collection System
+3. Change name of `FritzVisionSegmentationModel` to `FritzVisionSegmentationPredictor` for consistency with other FritzVision features.
+
 ## [5.3.4](https://github.com/fritzlabs/swift-framework/releases/tag/5.3.4)
 
 1. Removing bounding box annotation from pose recording.